clean up template stuff with anchors/aliases
I'd honestly prefer to be able to have templates use other templates
than use this feature of YAML that I've never super been into, but this
is the way.

It's clean-up in a way, sure. It also makes the "name prefix" stuff
easier, which I wanted so there's not always a space before the name.