James Acaster playlist
This was a pain to get working, because most things will return an
entire show if a person's in just one episode. And I don't want this
playlist to have all of BFQ, WILTY, and Taskmaster. I just want that
Acaster goodness (and weirdness).

Unfortunately, there are a lot of results where TMDb doesn't have the
IMDb external ID, so currently my playlist only has e.g. a few episodes
of Hypothetical instead of the whole thing.

That will be slowly fixed outside of this repo. Le sigh.