simplify the common template anchors
There doesn't have to be this separation between the common and
anchor-holder stuff. If defaults or conditionals are overridden at all,
they'll take precedence over what's from common (and be extended or not,
depending on use). And this just means I don't have to repeat all the
conditional stuff that is used but not changed.