fix top-of-list ordering _again_
It's not that + is special and different from !. It's now that longer
repeats go later.

I think Plex started separating sort titles on _ and comparing each
section. That would make sense with what I'm seeing:

- ++++ after +
- _+_ after __

Just going to go with the flow and what I see, and I'll adjust sort
titles as I appear to need to.